refactor(ivy): capture source span information in `PartialInterpreter`
Currently all source mapping information gets lost when
we statically interpret an expression that has been obtained
from source code.

This change modifies the `ResolvedValue` and related types
so that they hold both the original location information from
where that value was sourced.

In order to do this, `ResolvedValue` now has a `span` property that
is computed based on the span or node of the resolved value and its
`container` resolved value, which are passed in the constructor.