refactor(ivy): wrap `ResolvedValue` in an object to allow capturing metadata
Previously, the `PartialInterpreter` returned native types as interpreted values.

This change modifies the `ResolvedValue` and related types
so that they wrap the interpreted value in an object which in the future
could hold additional metadata about the nodes of the interpreted value,
such as the SourceFile or SourceSpan information.

`ResolvedValue` now has `value`. Moreover, if the resolved value is a
container, such as an array or a map, then the contained elements are also
`ResolvedValue` objects.

Code that uses `ResolvedValue` instances must be aware of this
wrapping of the interpreted values. Each `ResolvedValue` object
has an `unwrap()` method that extracts `value` (recursively
in the case of a container) that client code can use to get the raw
value.